refactor: re-define ripple builder code
create a fn that return a listener fn that have access to the target element when attached to the `mousedown` event
